Authorization + Migrations
Rewrote large sections of the blog to allow for parts of the site
to be hidden based on authentication. Now, all of the pages have
links to them (now logout & post creation is easier).

Also moved the migrations to a better place using postgres-simple-
migration. They are now run at start if they hadn't already been started.