SETUID compat & HTTPS redirection
Improved Blog.Daemonize so that it's more general and can be used
to spawn/detach a process in addition to daemonizing. This is so
that an HTTP to HTTPS redirection process can be started.

Also wrote in a handler to get rid of root privs once the HTTP port
has been bound to.