Caching improvements & Concat migrations
For assets, an Etag (for uncompressed data) would be calculated
using compressed data. Since FileCache stores compressed
representations, it also must store an Etag in memory. It saves
processing time and reports a correct value. In the future, it
may also store a compressed Etag.

After many iterations of adding features to the blog, the
migrations got out of hand and needed to be compressed into a
single file so as to reduce confusion later.